frysk.sys.Pid.* returns a ProcessIdentifier.
frysk-core/frysk/debuginfo/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestDebugInfoStackTrace.java: Update to match frysk.sys.Pid.

frysk-core/frysk/proc/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestHostGet.java: Update to match frysk.sys.Pid.
	* TestTaskTerminateObserver.java: Ditto.
	* TestTaskSyscallObserver.java: Ditto.

frysk-core/frysk/proc/live/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* LinuxPtraceHost.java: Update to match frysk.sys.Pid.
	* TestRefresh.java: Ditto.

frysk-core/frysk/stepping/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* TestStepping.java: Update to match frysk.sys.Pid.

frysk-core/frysk/testbed/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* ExecCommand.java: Update to match frysk.sys.Pid.
	* TestLib.java: Ditto.
	* SlaveOffspring.java: Ditto.
	* LegacyOffspring.java: Ditto.
	* FunitThreadsOffspring.java: Ditto.

frysk-sys/frysk/sys/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* Pid.java: Return a ProcessIdentifier.

frysk-sys/lib/dwfl/ChangeLog
2008-02-12  Andrew Cagney  <cagney@redhat.com>

	* Dwfl.java (Dwfl(ProcessIdentifier,String)): Replace
	Dwfl(int,String).

+++ frysk-core/frysk/debuginfo/TestDebugInfoStackTrace.java
@@ -98,7 +98,7 @@ public class TestDebugInfoStackTrace
     SynchronizedOffspring process
 	= new SynchronizedOffspring (Signal.USR1, new String[] {
 					 getExecPath ("funit-rt-looper"),
-					 Integer.toString(Pid.get ()),
+					 Integer.toString(Pid.get().intValue()),
 					 Integer.toString(Signal.USR1.intValue())
 				     });
     myTask = process.findTaskUsingRefresh(true);
@@ -216,7 +216,7 @@ public class TestDebugInfoStackTrace
     SynchronizedOffspring process
 	= new SynchronizedOffspring (Signal.USR1, new String[] {
 					 getExecPath ("funit-rt-threader"),
-					 Integer.toString(Pid.get ()),
+					 Integer.toString(Pid.get().intValue()),
 					 Integer.toString(Signal.USR1.intValue())
 				     });
     myTask = process.findTaskUsingRefresh(true);
@@ -266,7 +266,7 @@ public class TestDebugInfoStackTrace
   SynchronizedOffspring process
       = new SynchronizedOffspring (Signal.USR1, new String[] {
 				       getExecPath ("funit-rt-stepper"),
-				       Integer.toString(Pid.get()),
+				       Integer.toString(Pid.get().intValue()),
 				       Integer.toString(Signal.USR1.intValue())
 				   });
   myTask = process.findTaskUsingRefresh(true);
@@ -304,7 +304,7 @@ public class TestDebugInfoStackTrace
   SynchronizedOffspring process
       = new SynchronizedOffspring (Signal.USR1, new String[] {
 				       getExecPath ("funit-rt-stepper"),
-				       Integer.toString(Pid.get ()),
+				       Integer.toString(Pid.get().intValue()),
 				       Integer.toString(Signal.USR1.intValue())
 				   });
   myTask = process.findTaskUsingRefresh(true);
